+++ arviz/plots/autocorrplot.py
@@ -1,9 +1,9 @@
 """Autocorrelation plot of data."""
 import numpy as np
 
+from arviz import convert_to_dataset
+from arviz.stats.diagnostics import autocorr
 from .plot_utils import _scale_text, default_grid, make_label, xarray_var_iter, _create_axes_grid
-from ..utils import convert_to_dataset
-from ..stats.diagnostics import autocorr
 
 
 def autocorrplot(data, var_names=None, max_lag=100, combined=False, figsize=None, textsize=None):
