+++ arviz/plots/forestplot.py
@@ -5,10 +5,10 @@ from itertools import tee
 import numpy as np
 import matplotlib.pyplot as plt
 
-from ..stats.diagnostics import _get_neff, _get_rhat
-from ..stats import hpd
+from arviz import convert_to_dataset
+from arviz.stats import hpd
+from arviz.stats.diagnostics import _get_neff, _get_rhat
 from .plot_utils import _scale_text, xarray_var_iter, make_label
-from ..utils import convert_to_dataset
 from .kdeplot import _fast_kde
