+++ arviz/plots/jointplot.py
@@ -2,8 +2,8 @@
 import matplotlib.pyplot as plt
 from matplotlib.ticker import NullFormatter
 
+from arviz import convert_to_dataset
 from .kdeplot import kdeplot
-from ..utils import convert_to_dataset
 from .plot_utils import _scale_text, get_bins, xarray_var_iter, make_label, get_coords
