+++ arviz/stats/__init__.py
@@ -1,3 +1,4 @@
+# pylint: disable=wildcard-import
 """Statistical tests and diagnostics for ArviZ."""
-from .stats import bfmi, compare, hpd, loo, r2_score, summary, waic, psislw
-from .diagnostics import effective_n, gelman_rubin, geweke
+from .stats import *
+from .diagnostics import *
