+++ arviz/tests/test_plot_utils.py
