--- arviz/utils/__init__.py
@@ -1,5 +1,0 @@
-"""Utility module for ArviZ."""
-from .utils import load_data, load_arviz_data
-
-from .xarray_utils import (convert_to_dataset, convert_to_inference_data, pymc3_to_inference_data,
-                           pystan_to_inference_data)
