+++ examples/ppcplot.py
@@ -9,4 +9,4 @@ import arviz as az
 az.style.use('arviz-darkgrid')
 
 data = az.load_arviz_data('non_centered_eight')
-az.ppcplot(data, alpha=0.03, figsize=(12, 6), textsize=14);
+az.ppcplot(data, alpha=0.03, figsize=(12, 6), textsize=14)
